scripts updated:

+++ scripts/application_stop.sh
@@ -1,2 +1,5 @@
 echo "Stopping any existing node servers"
-pkill node
+if [ node ]
+then
+    pkill node
+fi
